reconcileInitialCNSState
